# openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Let's talk less noise",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/usage/checkpoint/end": {
            "put": {
                "tags": [
                    "usage"
                ],
                "summary": "Update Checkpoint",
                "operationId": "update_checkpoint_api_usage_checkpoint_end_put",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CheckpointEnd"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateOneResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/usage/attempt/end": {
            "put": {
                "tags": [
                    "usage"
                ],
                "summary": "Update Checkpoint",
                "operationId": "update_checkpoint_api_usage_attempt_end_put",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AttemptEnd"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateOneResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AttemptEnd": {
                "title": "AttemptEnd",
                "required": [
                    "attemptId",
                    "endTime",
                    "complete"
                ],
                "type": "object",
                "properties": {
                    "attemptId": {
                        "title": "Attemptid",
                        "type": "string"
                    },
                    "endTime": {
                        "title": "Endtime",
                        "type": "string",
                        "format": "date-time"
                    },
                    "complete": {
                        "title": "Complete",
                        "type": "boolean"
                    },
                    "failReason": {
                        "title": "Failreason",
                        "type": "string"
                    }
                },
                "example": {
                    "attemptId": "\u003Cmachine\u003E-\u003Cuuid\u003E",
                    "endTime": "2021-07-10T21:28:51.860777",
                    "complete": false,
                    "failReason": "Timeout"
                }
            },
...
            "CheckpointEnd": {
                "title": "CheckpointEnd",
                "required": [
                    "attemptId",
                    "description",
                    "end"
                ],
                "type": "object",
                "properties": {
                    "attemptId": {
                        "title": "Attemptid",
                        "type": "string"
                    },
                    "description": {
                        "title": "Description",
                        "type": "string"
                    },
                    "end": {
                        "title": "End",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "example": {
                    "attemptId": "\u003Cmachine\u003E-\u003Cuuid\u003E",
                    "description": "noiseCollation",
                    "end": "2021-07-10T21:28:51.860777"
                }
            },
...
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
...
            "UpdateOneResponse": {
                "title": "UpdateOneResponse",
                "required": [
                    "acknowledged",
                    "matchedCount",
                    "modifiedCount"
                ],
                "type": "object",
                "properties": {
                    "acknowledged": {
                        "title": "Acknowledged",
                        "type": "boolean"
                    },
                    "matchedCount": {
                        "title": "Matchedcount",
                        "type": "integer"
                    },
                    "modifiedCount": {
                        "title": "Modifiedcount",
                        "type": "integer"
                    }
                },
                "example": {
                    "matchedCount": 1,
                    "modifiedCount": 1,
                    "acknowledged": true
                }
            },
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            }
        }
    }
}